feat: implement shipping service and refactor some parts
Replaced float with double for product prices, customer balances, and cart subtotals for improved precision. Added ShippingService to handle shipping of shippable items and integrated it into the checkout process. Enhanced Cart with resetCart and improved addProduct validation. Updated Main to reflect new logic and test cases.

+++ fawry-intern-challenge/src/products/Product.java
@@ -3,10 +3,10 @@ package products;
 public abstract class Product {
 
     protected String name;
-    protected float price;
+    protected double price;
     protected int  quantity;
 
-    public Product(String name, float price, int quantity) {
+    public Product(String name, double price, int quantity) {
         this.name = name;
         this.price = price;
         this.quantity = quantity;
@@ -20,11 +20,11 @@ public abstract class Product {
         this.name = name;
     }
 
-    public float getPrice() {
+    public double getPrice() {
         return price;
     }
 
-    public void setPrice(float price) {
+    public void setPrice(double price) {
         this.price = price;
     }
